use -j4 for make update reset characteristic to use uint8 init mesh earlier

+++ include/services/cs_GeneralService.h
@@ -70,5 +70,5 @@ private:
 	 *
 	 * Resets device
 	 */
-	BLEpp::Characteristic<int32_t>* _resetCharacteristic;
+	BLEpp::Characteristic<uint8_t>* _resetCharacteristic;
 };
